Address review comments by takoverflow - part 1
